Rename QuickStartSettings.cs to QuickstartSettings.cs
